refactor(ui): rename dismissGuard to useDismissGuard, drop unused test shim
Match the ui/ hook file naming convention (useSettingsLayout, useToast)
and remove the globalThis.React assignment from the test — the module
has no JSX, and the suite passes without it.
